@@ -15,11 +15,9 @@ License: [BSD](https://opensource.org/licenses/bsd-license.php)
 
 """
 
-from __future__ import absolute_import
-from __future__ import unicode_literals
 from . import Extension
 from ..blockprocessors import BlockProcessor
-from ..util import etree
+import xml.etree.ElementTree as etree
 import re
 PIPE_NONE = 0
 PIPE_LEFT = 1
@@ -35,7 +33,7 @@ class TableProcessor(BlockProcessor):
     def __init__(self, parser):
         self.border = False
         self.separator = ''
-        super(TableProcessor, self).__init__(parser)
+        super().__init__(parser)
 
     def test(self, parent, block):
         """
