@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Python Markdown
 
@@ -25,7 +24,6 @@ Python-Markdown Regression Tests
 Tests of the various APIs with the python markdown lib.
 """
 
-from __future__ import unicode_literals
 import unittest
 import sys
 import os
@@ -36,15 +34,8 @@ from logging import DEBUG, WARNING, CRITICAL
 import yaml
 import tempfile
 from io import BytesIO
+import xml.etree.ElementTree as etree
 from xml.etree.ElementTree import ProcessingInstruction
-
-
-PY3 = sys.version_info[0] == 3
-
-
-if not PY3:
-    def bytes(string, encoding):
-        return string.encode(encoding)
 
 
 class TestMarkdownBasics(unittest.TestCase):
@@ -132,7 +123,7 @@ class TestBlockParser(unittest.TestCase):
 
     def testParseChunk(self):
         """ Test BlockParser.parseChunk. """
-        root = markdown.util.etree.Element("div")
+        root = etree.Element("div")
         text = 'foo'
         self.parser.parseChunk(root, text)
         self.assertEqual(
@@ -144,8 +135,8 @@ class TestBlockParser(unittest.TestCase):
         """ Test BlockParser.parseDocument. """
         lines = ['#foo', '', 'bar', '', '    baz']
         tree = self.parser.parseDocument(lines)
-        self.assertIsInstance(tree, markdown.util.etree.ElementTree)
-        self.assertIs(markdown.util.etree.iselement(tree.getroot()), True)
+        self.assertIsInstance(tree, etree.ElementTree)
+        self.assertIs(etree.iselement(tree.getroot()), True)
         self.assertEqual(
             markdown.serializers.to_xhtml_string(tree.getroot()),
             "<div><h1>foo</h1><p>bar</p><pre><code>baz\n</code></pre></div>"
@@ -220,7 +211,7 @@ class TestHtmlStash(unittest.TestCase):
         self.assertEqual(self.stash.rawHtmlBlocks, [])
 
 
-class Item(object):
+class Item:
     """ A dummy Registry item object for testing. """
     def __init__(self, data):
         self.data = data
@@ -346,7 +337,7 @@ class RegistryTests(unittest.TestCase):
     def testRegistryDelItem(self):
         r = markdown.util.Registry()
         r.register(Item('a'), 'a', 20)
-        with self.assertRaises(TypeError):
+        with self.assertRaises(KeyError):
             del r[0]
         # TODO: restore this when deprecated __del__ is removed.
         # with self.assertRaises(TypeError):
@@ -361,7 +352,7 @@ class RegistryTests(unittest.TestCase):
             self.assertEqual(list(r), ['a', 'c'])
             del r['a']
             self.assertEqual(list(r), ['c'])
-            with self.assertRaises(TypeError):
+            with self.assertRaises(KeyError):
                 del r['badname']
             del r['c']
             self.assertEqual(list(r), [])
@@ -444,12 +435,6 @@ class TestErrors(unittest.TestCase):
         # Reset warning behavior back to default
         warnings.simplefilter('default')
 
-    def testNonUnicodeSource(self):
-        """ Test falure on non-unicode source text. """
-        if not PY3:
-            source = "foo".encode('utf-16')
-            self.assertRaises(UnicodeDecodeError, markdown.markdown, source)
-
     def testBadOutputFormat(self):
         """ Test failure on bad output_format. """
         self.assertRaises(KeyError, markdown.Markdown, output_format='invalid')
@@ -498,15 +483,11 @@ class testETreeComments(unittest.TestCase):
 
     def setUp(self):
         # Create comment node
-        self.comment = markdown.util.etree.Comment('foo')
-        if hasattr(markdown.util.etree, 'test_comment'):
-            self.test_comment = markdown.util.etree.test_comment
-        else:
-            self.test_comment = markdown.util.etree.Comment
+        self.comment = etree.Comment('foo')
 
     def testCommentIsComment(self):
         """ Test that an ElementTree Comment passes the `is Comment` test. """
-        self.assertIs(self.comment.tag, markdown.util.etree.test_comment)
+        self.assertIs(self.comment.tag, etree.Comment)
 
     def testCommentIsBlockLevel(self):
         """ Test that an ElementTree Comment is recognized as BlockLevel. """
@@ -537,8 +518,8 @@ class testElementTailTests(unittest.TestCase):
 
     def testBrTailNoNewline(self):
         """ Test that last <br> in tree has a new line tail """
-        root = markdown.util.etree.Element('root')
-        br = markdown.util.etree.SubElement(root, 'br')
+        root = etree.Element('root')
+        br = etree.SubElement(root, 'br')
         self.assertEqual(br.tail, None)
         self.pretty.run(root)
         self.assertEqual(br.tail, "\n")
@@ -549,15 +530,15 @@ class testSerializers(unittest.TestCase):
 
     def testHtml(self):
         """ Test HTML serialization. """
-        el = markdown.util.etree.Element('div')
+        el = etree.Element('div')
         el.set('id', 'foo<&">')
-        p = markdown.util.etree.SubElement(el, 'p')
+        p = etree.SubElement(el, 'p')
         p.text = 'foo <&escaped>'
         p.set('hidden', 'hidden')
-        markdown.util.etree.SubElement(el, 'hr')
-        non_element = markdown.util.etree.SubElement(el, None)
+        etree.SubElement(el, 'hr')
+        non_element = etree.SubElement(el, None)
         non_element.text = 'non-element text'
-        script = markdown.util.etree.SubElement(non_element, 'script')
+        script = etree.SubElement(non_element, 'script')
         script.text = '<&"test\nescaping">'
         el.tail = "tail text"
         self.assertEqual(
@@ -572,15 +553,15 @@ class testSerializers(unittest.TestCase):
 
     def testXhtml(self):
         """" Test XHTML serialization. """
-        el = markdown.util.etree.Element('div')
+        el = etree.Element('div')
         el.set('id', 'foo<&">')
-        p = markdown.util.etree.SubElement(el, 'p')
+        p = etree.SubElement(el, 'p')
         p.text = 'foo<&escaped>'
         p.set('hidden', 'hidden')
-        markdown.util.etree.SubElement(el, 'hr')
-        non_element = markdown.util.etree.SubElement(el, None)
+        etree.SubElement(el, 'hr')
+        non_element = etree.SubElement(el, None)
         non_element.text = 'non-element text'
-        script = markdown.util.etree.SubElement(non_element, 'script')
+        script = etree.SubElement(non_element, 'script')
         script.text = '<&"test\nescaping">'
         el.tail = "tail text"
         self.assertEqual(
@@ -595,11 +576,11 @@ class testSerializers(unittest.TestCase):
 
     def testMixedCaseTags(self):
         """" Test preservation of tag case. """
-        el = markdown.util.etree.Element('MixedCase')
+        el = etree.Element('MixedCase')
         el.text = 'not valid '
-        em = markdown.util.etree.SubElement(el, 'EMPHASIS')
+        em = etree.SubElement(el, 'EMPHASIS')
         em.text = 'html'
-        markdown.util.etree.SubElement(el, 'HR')
+        etree.SubElement(el, 'HR')
         self.assertEqual(
             markdown.serializers.to_xhtml_string(el),
             '<MixedCase>not valid <EMPHASIS>html</EMPHASIS><HR /></MixedCase>'
@@ -616,17 +597,17 @@ class testSerializers(unittest.TestCase):
 
     def testQNameTag(self):
         """ Test serialization of QName tag. """
-        div = markdown.util.etree.Element('div')
-        qname = markdown.util.etree.QName('http://www.w3.org/1998/Math/MathML', 'math')
-        math = markdown.util.etree.SubElement(div, qname)
+        div = etree.Element('div')
+        qname = etree.QName('http://www.w3.org/1998/Math/MathML', 'math')
+        math = etree.SubElement(div, qname)
         math.set('display', 'block')
-        sem = markdown.util.etree.SubElement(math, 'semantics')
-        msup = markdown.util.etree.SubElement(sem, 'msup')
-        mi = markdown.util.etree.SubElement(msup, 'mi')
+        sem = etree.SubElement(math, 'semantics')
+        msup = etree.SubElement(sem, 'msup')
+        mi = etree.SubElement(msup, 'mi')
         mi.text = 'x'
-        mn = markdown.util.etree.SubElement(msup, 'mn')
+        mn = etree.SubElement(msup, 'mn')
         mn.text = '2'
-        ann = markdown.util.etree.SubElement(sem, 'annotations')
+        ann = etree.SubElement(sem, 'annotations')
         ann.text = 'x^2'
         self.assertEqual(
             markdown.serializers.to_xhtml_string(div),
@@ -645,8 +626,8 @@ class testSerializers(unittest.TestCase):
 
     def testQNameAttribute(self):
         """ Test serialization of QName attribute. """
-        div = markdown.util.etree.Element('div')
-        div.set(markdown.util.etree.QName('foo'), markdown.util.etree.QName('bar'))
+        div = etree.Element('div')
+        div.set(etree.QName('foo'), etree.QName('bar'))
         self.assertEqual(
             markdown.serializers.to_xhtml_string(div),
             '<div foo="bar"></div>'
@@ -654,14 +635,14 @@ class testSerializers(unittest.TestCase):
 
     def testBadQNameTag(self):
         """ Test serialization of QName with no tag. """
-        qname = markdown.util.etree.QName('http://www.w3.org/1998/Math/MathML')
-        el = markdown.util.etree.Element(qname)
+        qname = etree.QName('http://www.w3.org/1998/Math/MathML')
+        el = etree.Element(qname)
         self.assertRaises(ValueError, markdown.serializers.to_xhtml_string, el)
 
     def testQNameEscaping(self):
         """ Test QName escaping. """
-        qname = markdown.util.etree.QName('<&"test\nescaping">', 'div')
-        el = markdown.util.etree.Element(qname)
+        qname = etree.QName('<&"test\nescaping">', 'div')
+        el = etree.Element(qname)
         self.assertEqual(
             markdown.serializers.to_xhtml_string(el),
             '<div xmlns="&lt;&amp;&quot;test&#10;escaping&quot;&gt;"></div>'
@@ -669,8 +650,8 @@ class testSerializers(unittest.TestCase):
 
     def testQNamePreEscaping(self):
         """ Test QName that is already partially escaped. """
-        qname = markdown.util.etree.QName('&lt;&amp;"test&#10;escaping"&gt;', 'div')
-        el = markdown.util.etree.Element(qname)
+        qname = etree.QName('&lt;&amp;"test&#10;escaping"&gt;', 'div')
+        el = etree.Element(qname)
         self.assertEqual(
             markdown.serializers.to_xhtml_string(el),
             '<div xmlns="&lt;&amp;&quot;test&#10;escaping&quot;&gt;"></div>'
@@ -718,8 +699,8 @@ class testAtomicString(unittest.TestCase):
 
     def testString(self):
         """ Test that a regular string is parsed. """
-        tree = markdown.util.etree.Element('div')
-        p = markdown.util.etree.SubElement(tree, 'p')
+        tree = etree.Element('div')
+        p = etree.SubElement(tree, 'p')
         p.text = 'some *text*'
         new = self.inlineprocessor.run(tree)
         self.assertEqual(
@@ -729,8 +710,8 @@ class testAtomicString(unittest.TestCase):
 
     def testSimpleAtomicString(self):
         """ Test that a simple AtomicString is not parsed. """
-        tree = markdown.util.etree.Element('div')
-        p = markdown.util.etree.SubElement(tree, 'p')
+        tree = etree.Element('div')
+        p = etree.SubElement(tree, 'p')
         p.text = markdown.util.AtomicString('some *text*')
         new = self.inlineprocessor.run(tree)
         self.assertEqual(
@@ -740,14 +721,14 @@ class testAtomicString(unittest.TestCase):
 
     def testNestedAtomicString(self):
         """ Test that a nested AtomicString is not parsed. """
-        tree = markdown.util.etree.Element('div')
-        p = markdown.util.etree.SubElement(tree, 'p')
+        tree = etree.Element('div')
+        p = etree.SubElement(tree, 'p')
         p.text = markdown.util.AtomicString('*some* ')
-        span1 = markdown.util.etree.SubElement(p, 'span')
+        span1 = etree.SubElement(p, 'span')
         span1.text = markdown.util.AtomicString('*more* ')
-        span2 = markdown.util.etree.SubElement(span1, 'span')
+        span2 = etree.SubElement(span1, 'span')
         span2.text = markdown.util.AtomicString('*text* ')
-        span3 = markdown.util.etree.SubElement(span2, 'span')
+        span3 = etree.SubElement(span2, 'span')
         span3.text = markdown.util.AtomicString('*here*')
         span3.tail = markdown.util.AtomicString(' *to*')
         span2.tail = markdown.util.AtomicString(' *test*')
@@ -865,7 +846,7 @@ class TestCliOptionParsing(unittest.TestCase):
 
     def create_config_file(self, config):
         """ Helper to create temp config files. """
-        if not isinstance(config, markdown.util.string_type):
+        if not isinstance(config, str):
             # convert to string
             config = yaml.dump(config)
         fd, self.tempfile = tempfile.mkstemp('.yml')
@@ -964,7 +945,7 @@ class TestAncestorExclusion(unittest.TestCase):
 
         def handleMatch(self, m, data):
             """ Handle match. """
-            el = markdown.util.etree.Element(self.tag)
+            el = etree.Element(self.tag)
             el.text = m.group(2)
             return el, m.start(0), m.end(0)
 
