@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Python Markdown
 
@@ -20,9 +19,7 @@ Copyright 2004 Manfred Stienstra (the original version)
 License: BSD (see LICENSE.md for details).
 """
 
-from __future__ import absolute_import
 import os
-import io
 import unittest
 import textwrap
 from . import markdown
@@ -108,9 +105,9 @@ class LegacyTestMeta(type):
 
         def generate_test(infile, outfile, normalize, kwargs):
             def test(self):
-                with io.open(infile, encoding="utf-8") as f:
+                with open(infile, encoding="utf-8") as f:
                     input = f.read()
-                with io.open(outfile, encoding="utf-8") as f:
+                with open(outfile, encoding="utf-8") as f:
                     # Normalize line endings
                     # (on Windows, git may have altered line endings).
                     expected = f.read().replace("\r\n", "\n")
@@ -150,13 +147,7 @@ class LegacyTestMeta(type):
         return type.__new__(cls, name, bases, dct)
 
 
-# Define LegacyTestCase class with metaclass in Py2 & Py3 compatible way.
-# See https://stackoverflow.com/a/38668373/866026
-# TODO: If/when py2 support is dropped change to:
-# class LegacyTestCase(unittest.Testcase, metaclass=LegacyTestMeta)
-
-
-class LegacyTestCase(LegacyTestMeta('LegacyTestCase', (unittest.TestCase,), {'__slots__': ()})):
+class LegacyTestCase(unittest.TestCase, metaclass=LegacyTestMeta):
     """
     A `unittest.TestCase` subclass for running Markdown's legacy file-based tests.
 
@@ -176,7 +167,7 @@ class LegacyTestCase(LegacyTestMeta('LegacyTestCase', (unittest.TestCase,), {'__
                     arguments for all test files in the directory.
 
     In addition, properties can be defined for each individual set of test files within
-    the directory. The property should be given the name of the file wihtout the file
+    the directory. The property should be given the name of the file without the file
     extension. Any spaces and dashes in the filename should be replaced with
     underscores. The value of the property should be a `Kwargs` instance which
     contains the keyword arguments that should be passed to `Markdown` for that
