@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Python Markdown
 
@@ -20,8 +19,7 @@ Copyright 2004 Manfred Stienstra (the original version)
 License: BSD (see LICENSE.md for details).
 """
 
-from __future__ import unicode_literals
-from __future__ import absolute_import
+import xml.etree.ElementTree as etree
 from . import util
 from . import inlinepatterns
 
@@ -37,7 +35,7 @@ def build_treeprocessors(md, **kwargs):
 def isString(s):
     """ Check if it's string """
     if not isinstance(s, util.AtomicString):
-        return isinstance(s, util.string_type)
+        return isinstance(s, str)
     return False
 
 
@@ -310,12 +308,12 @@ class InlineProcessor(Treeprocessor):
         placeholder = self.__stashNode(node, pattern.type())
 
         if new_style:
-            return "%s%s%s" % (data[:start],
-                               placeholder, data[end:]), True, 0
+            return "{}{}{}".format(data[:start],
+                                   placeholder, data[end:]), True, 0
         else:  # pragma: no cover
-            return "%s%s%s%s" % (leftData,
-                                 match.group(1),
-                                 placeholder, match.groups()[-1]), True, 0
+            return "{}{}{}{}".format(leftData,
+                                     match.group(1),
+                                     placeholder, match.groups()[-1]), True, 0
 
     def __build_ancestors(self, parent, parents):
         """Build the ancestor list."""
@@ -351,7 +349,7 @@ class InlineProcessor(Treeprocessor):
         # to ensure we don't have the user accidentally change it on us.
         tree_parents = [] if ancestors is None else ancestors[:]
 
-        self.parent_map = dict((c, p) for p in tree.iter() for c in p)
+        self.parent_map = {c: p for p in tree.iter() for c in p}
         stack = [(tree, tree_parents)]
 
         while stack:
@@ -378,7 +376,7 @@ class InlineProcessor(Treeprocessor):
                     self.ancestors.pop()
                 if child.tail:
                     tail = self.__handleInline(child.tail)
-                    dumby = util.etree.Element('d')
+                    dumby = etree.Element('d')
                     child.tail = None
                     tailResult = self.__processPlaceholders(tail, dumby, False)
                     if dumby.tail:
