@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Python Markdown
 
@@ -83,4 +82,35 @@ class TestNotEmphasis(TestCase):
         self.assertMarkdownRenders(
             '_ bar _',
             '<p>_ bar _</p>'
+        )
+
+    def test_complex_emphasis_asterisk(self):
+        self.assertMarkdownRenders(
+            'This is text **bold *italic bold*** with more text',
+            '<p>This is text <strong>bold <em>italic bold</em></strong> with more text</p>'
+        )
+
+    def test_complex_emphasis_asterisk_mid_word(self):
+        self.assertMarkdownRenders(
+            'This is text **bold*italic bold*** with more text',
+            '<p>This is text <strong>bold<em>italic bold</em></strong> with more text</p>'
+        )
+
+    def test_complex_emphasis_smart_underscore(self):
+        self.assertMarkdownRenders(
+            'This is text __bold _italic bold___ with more text',
+            '<p>This is text <strong>bold <em>italic bold</em></strong> with more text</p>'
+        )
+
+    def test_complex_emphasis_smart_underscore_mid_word(self):
+        self.assertMarkdownRenders(
+            'This is text __bold_italic bold___ with more text',
+            '<p>This is text __bold_italic bold___ with more text</p>'
+        )
+
+    def test_nested_emphasis(self):
+
+        self.assertMarkdownRenders(
+            'This text is **bold *italic* *italic* bold**',
+            '<p>This text is <strong>bold <em>italic</em> <em>italic</em> bold</strong></p>'
         )
